Let the fp32 output heads run under a wholesale bf16/half cast
FinalLayer3D and PixelDiTLastLayer force the output projection to fp32
(norm on x.float(), autocast disabled) for numerical stability. That works
under bf16_mixed autocast, where weights stay fp32, but crashes when the
model is cast wholesale with model.bfloat16() / .half(): the fp32 input then
meets a bf16 linear weight and F.linear raises "mat1 and mat2 must have the
same dtype". Pure-half casting is a plausible deployment given the models'
MetaData advertises bf16.

Match the linear input to the weight dtype. Under the common fp32-weight
path the cast is a no-op, so existing numerics (and goldens) are unchanged.

Add a regression test that casts DiT3D and PixelDiT to bf16 and runs forward
(the PixelDiT case also exercises the bf16 DepthwiseConv bilinear_dw path).

+++ physicsnemo/experimental/models/strata/layers.py
@@ -510,7 +510,10 @@ class FinalLayer3D(nn.Module):
         self, x: Float[torch.Tensor, "batch tokens hidden_size"]
     ) -> Float[torch.Tensor, "batch tokens patch_pixels"]:
         # Force the output head to fp32 regardless of any outer autocast context.
+        # Match the linear input to the weight dtype so the model also works when
+        # cast wholesale to bf16/half (model.bfloat16()); under fp32 weights (the
+        # common autocast path) the cast is a no-op and the result is unchanged.
         with torch.autocast(device_type=x.device.type, enabled=False):
             x = self.norm(x.float())
-            x = self.linear(x)
+            x = self.linear(x.to(self.linear.weight.dtype))
         return x
